refactor: mechanical code-quality wins (7 of 16 findings) (#771)
* refactor: dedup test fixtures, split process/styles/sound modules

Cleanup-pass findings #10, #16, #12, #13 from the code-quality scan.
No behavior changes; all tests still pass.

- tests/common/mod.rs: shared setup_temp_home/set_temp_home, replacing
  eight identical definitions across integration test files.
- src/process/: kill_with_fallback + process_exists moved to the shared
  mod.rs (using portable nix kill(None) for existence). Each platform
  module now only exposes collect_pid_tree (Linux via /proc, macOS via
  ps) plus get_foreground_pid.
- src/tui/styles/: split the 857-line file into palette.rs (RGB->xterm
  palette mapping), themes.rs (Theme struct + built-in constructors +
  hex_color serde helper), and mod.rs (loading/discovery glue).
- src/sound/: split the 717-line file into config.rs (SoundConfig +
  volume helpers), discovery.rs (sounds dir probing), bundled.rs
  (GitHub installer), playback.rs (afplay/paplay/aplay dispatch), and
  mod.rs (transition-to-sound glue).


* refactor: collapse container-runtime wrappers and paired-terminal twins

Cleanup-pass findings #15 and #14 from the code-quality scan.
No behavior changes.

- src/containers/: Docker and AppleContainer were near-identical thin
  wrappers around RuntimeBase. Collapsed into a single ContainerRuntime
  { base: RuntimeBase, kind: RuntimeKind } that dispatches the four
  genuinely runtime-specific operations (existence probe, running-state
  probe, exec-command formatting, batch status query) on an enum.
  Eliminates ~390 lines of wrapper boilerplate and drops the
  enum_dispatch dependency.
- src/tmux/terminal_session.rs: TerminalSession and
  ContainerTerminalSession shared ~90% of their bodies. Extracted a
  private PairedTerminal struct with a TerminalKind discriminant that
  controls prefix and error-message label; the two public types are
  thin wrappers that preserve the external API.


* refactor(git): split mod.rs into remote.rs and worktree.rs

Cleanup-pass finding #5 from the code-quality scan. No behavior changes.

Before: src/git/mod.rs was 1,797 lines hosting three distinct concerns.
After:
  - src/git/mod.rs (35 lines) — module declarations, re-exports, and the
    shared pub(crate) open_repo_at helper used by sibling submodules.
  - src/git/remote.rs (169 lines) — clone_repo, parse_owner_from_remote_url,
    get_remote_owner, and their parse tests.
  - src/git/worktree.rs (1,624 lines) — GitWorktree lifecycle methods,
    branch ops, template path expansion, and all the worktree tests.

Public API (crate::git::{GitWorktree, WorktreeEntry, clone_repo,
get_remote_owner}) is preserved, so downstream consumers in cli/,
session/, and server/ are unchanged.


* refactor(cli,tui): cli/output formatter + shared checkbox helper

Cleanup-pass findings #11 and #9 from the code-quality scan. No behavior
changes; all 1079 unit tests pass.

- src/cli/output.rs: tiny print_json helper. Six callers in
  cli/{group,list,session}.rs now go through it instead of inlining
  println!("{}", serde_json::to_string_pretty(&x)?). Single place to add
  a global --output-format flag later.

- src/tui/components/checkbox.rs: shared checkbox_line builder + a
  CheckboxStyle config so each dialog keeps its own focus-highlight
  taste (delete_options uses bold+accent; group_delete_options
  underlines the row). delete_options.rs and group_delete_options.rs
  now call into it for ~7 checkbox renders, removing the copy-pasted
  glyph + style + indent code.


* refactor(tui): collapse dialog render dispatch with a macro

Cleanup-pass finding #7 from the code-quality scan. Modest win.

Replaces 13 near-identical `if let Some(dialog) = &self.X { dialog.render(...) }`
blocks in HomeView::render_dialogs with a single `render_dialogs!`
macro invocation that takes the list of dialog field names. The
serve_dialog dispatch stays separate because of the `#[cfg(feature =
"serve")]` gate.

Why a macro instead of the trait-based registry the scan suggested:
each dialog returns a different DialogResult<T> (DeleteOptions,
GroupDeleteOptions, String, ()), so the *input* dispatch can't be
unified without erasing types and re-matching downstream. The render
path is the only piece that's truly uniform, and a macro flattens it
without trait boilerplate.


* refactor(tui): shared dialog button-row helper

Cleanup-pass finding #8 from the code-quality scan. Partial: extracts
the most copy-pasted piece (the centered Yes/No button row) without
forcing the dialogs into a single ModalForm abstraction — they have
genuinely different shapes (TextArea vs checkbox vs picker) and the
shared bits don't justify a parameterized super-class.

- src/tui/components/buttons.rs: render_buttons(&[DialogButton]) plus a
  render_yes_no convenience for the common destructive-confirm row.
- confirm.rs and delete_options.rs route their button rendering through
  the helper, removing two copies of the [Yes] / [No] glyph/style logic.

Tests: all 1079 unit tests still pass.


* chore: clippy cleanup (modulo_one, useless vec!, unused mut)

- src/tui/status_poller.rs: cycle % TIER_HOT triggered modulo_one because
  TIER_HOT is 1 by design. Replace trivially-true modulo checks with one
  assert_eq!(TIER_HOT, 1) guard — the real invariant the test was trying
  to express.
- src/tui/home/tests.rs: drop unnecessary mut on a test-local Instance.
- tests/config_wiring.rs: vec!["..."] -> ["..."] in two spots where the
  allocation was useless.

cargo clippy --all-targets [--features serve] clean both ways.


* refactor: small review follow-ups

- src/cli/output.rs: print_json streams via to_writer_pretty against a
  locked stdout instead of buffering the entire JSON payload through
  to_string_pretty. Avoids doubling peak memory for large `aoe list --json`
  outputs.
- src/tui/components/buttons.rs: render_buttons + DialogButton were dead
  (only render_yes_no ever called render_buttons, and nothing outside
  this file called render_buttons directly). Inlined render_yes_no
  directly, removing ~30 lines of unused indirection. Also dropped a
  doc-comment claim that the helpers were "used by editor dialogs"
  (they weren't).
- tests/repo_config.rs: dropped the `set_temp_home as setup_temp_home`
  alias that hid a different signature from the other 7 callers. Now
  calls set_temp_home directly.

All 1213 lib tests + every integration suite still pass.


* style: cargo fmt (pub use runtime ordering)

* refactor: review follow-ups round 2

Three changes from review of PR #771.

- src/tui/components/checkbox.rs: CheckboxStyle had two overlapping
  underline flags (focused_label_underlined, focused_row_underlined)
  that could in principle both be true. Collapsed to
  focused_underline: Option<UnderlineScope::{Label,Row}>, which makes
  "label vs row" mutually exclusive at the type level and removes
  duplicate underline-the-label logic from the render path.

- src/process/linux.rs: collect_descendants used to re-scan /proc once
  per node in the tree (O(children x /proc scan)). Refactored to the
  same shape macos.rs already uses: build a parent->children map in one
  /proc pass, then DFS the map. Added the map-descent tests that were
  already covering the macOS path.

- src/git/mod.rs: doc comment listed diff/cleanup/template as new
  submodules from the split. They predate it. Clarified that only
  remote and worktree were extracted.

cargo clippy --all-targets --features serve: clean (the two remaining
warnings are in server/api.rs and fixed in the stacked PR).
cargo test --lib --features serve: 1217 passed.


---------
