+++ src/cli/mod.rs
@@ -5,6 +5,7 @@ pub mod definition;
 pub mod group;
 pub mod init;
 pub mod list;
+pub mod output;
 pub mod profile;
 pub mod remove;
 pub mod send;
