+++ src/process/linux.rs
@@ -1,81 +1,59 @@
 //! Linux-specific process utilities
 
+use std::collections::HashMap;
 use std::fs;
 use std::path::Path;
 
-use nix::sys::signal::{kill, Signal};
-use nix::unistd::Pid;
-use tracing::debug;
-
-/// Kill a process and all its descendants
-/// Uses SIGTERM first, then SIGKILL after a short delay for stragglers
-pub fn kill_process_tree(pid: u32) {
-    // Collect all descendant PIDs first (children, grandchildren, etc.)
-    let mut pids_to_kill = vec![pid];
-    collect_descendants(pid, &mut pids_to_kill);
-
-    debug!(
-        pid,
-        descendants = ?pids_to_kill,
-        "Killing process tree"
-    );
-
-    // Kill in reverse order (children first, then parent) with SIGTERM
-    for &p in pids_to_kill.iter().rev() {
-        let _ = kill(Pid::from_raw(p as i32), Signal::SIGTERM);
-    }
-
-    // Brief pause to let processes handle SIGTERM gracefully
-    std::thread::sleep(std::time::Duration::from_millis(100));
-
-    // SIGKILL any survivors
-    for &p in pids_to_kill.iter().rev() {
-        if process_exists(p) {
-            debug!(pid = p, "Process survived SIGTERM, sending SIGKILL");
-            let _ = kill(Pid::from_raw(p as i32), Signal::SIGKILL);
-        }
-    }
+/// Collect `pid` and every descendant by walking `/proc` once to build a
+/// parent -> children map, then descending it. One `/proc` scan regardless of
+/// tree depth.
+pub(super) fn collect_pid_tree(pid: u32) -> Vec<u32> {
+    let children_map = build_children_map();
+    let mut pids = vec![pid];
+    collect_descendants_from_map(pid, &children_map, &mut pids);
+    pids
 }
 
-/// Recursively collect all descendant PIDs of a process
-fn collect_descendants(pid: u32, pids: &mut Vec<u32>) {
+/// Scan `/proc` once and group every live PID by its parent.
+fn build_children_map() -> HashMap<u32, Vec<u32>> {
+    let mut children_map: HashMap<u32, Vec<u32>> = HashMap::new();
     let proc_dir = Path::new("/proc");
-    if !proc_dir.exists() {
-        return;
-    }
-
     let Ok(entries) = fs::read_dir(proc_dir) else {
-        return;
+        return children_map;
     };
 
     for entry in entries.flatten() {
         let name = entry.file_name();
         let name_str = name.to_string_lossy();
 
-        // Skip non-numeric entries
         let Ok(child_pid) = name_str.parse::<u32>() else {
             continue;
         };
 
-        // Read the process's parent PID
         let stat_path = entry.path().join("stat");
         let Ok(content) = fs::read_to_string(&stat_path) else {
             continue;
         };
 
         if let Some(ppid) = parse_stat_field(&content, 3) {
-            if ppid as u32 == pid {
-                pids.push(child_pid);
-                // Recurse to find grandchildren
-                collect_descendants(child_pid, pids);
-            }
+            children_map.entry(ppid as u32).or_default().push(child_pid);
         }
     }
+
+    children_map
 }
 
-/// Check if a process still exists
-fn process_exists(pid: u32) -> bool {
-    Path::new(&format!("/proc/{}", pid)).exists()
+fn collect_descendants_from_map(
+    pid: u32,
+    children_map: &HashMap<u32, Vec<u32>>,
+    pids: &mut Vec<u32>,
+) {
+    if let Some(children) = children_map.get(&pid) {
+        for &child_pid in children {
+            pids.push(child_pid);
+            collect_descendants_from_map(child_pid, children_map, pids);
+        }
+    }
 }
 
 /// Get the foreground process group leader for a shell PID
@@ -158,5 +136,57 @@ mod tests {
         assert_eq!(parse_stat_field(stat, 3), Some(1233)); // ppid
         assert_eq!(parse_stat_field(stat, 4), Some(1234)); // pgrp
         assert_eq!(parse_stat_field(stat, 7), Some(1234)); // tpgid
+    }
+
+    #[test]
+    fn test_collect_descendants_from_map_empty() {
+        let children_map = HashMap::new();
+        let mut pids = vec![100];
+        collect_descendants_from_map(100, &children_map, &mut pids);
+        assert_eq!(pids, vec![100]);
+    }
+
+    #[test]
+    fn test_collect_descendants_from_map_nested() {
+        // Tree: 100 -> 101 -> 102 -> 103
+        let mut children_map = HashMap::new();
+        children_map.insert(100, vec![101]);
+        children_map.insert(101, vec![102]);
+        children_map.insert(102, vec![103]);
+
+        let mut pids = vec![100];
+        collect_descendants_from_map(100, &children_map, &mut pids);
+        assert_eq!(pids, vec![100, 101, 102, 103]);
+    }
+
+    #[test]
+    fn test_collect_descendants_from_map_branching() {
+        // Tree: 100 -> [101, 102], 101 -> [103, 104], 102 -> [105]
+        let mut children_map = HashMap::new();
+        children_map.insert(100, vec![101, 102]);
+        children_map.insert(101, vec![103, 104]);
+        children_map.insert(102, vec![105]);
+
+        let mut pids = vec![100];
+        collect_descendants_from_map(100, &children_map, &mut pids);
+
+        assert!(pids.contains(&100));
+        assert!(pids.contains(&101));
+        assert!(pids.contains(&102));
+        assert!(pids.contains(&103));
+        assert!(pids.contains(&104));
+        assert!(pids.contains(&105));
+        assert_eq!(pids.len(), 6);
+    }
+
+    #[test]
+    fn test_collect_descendants_unrelated_processes() {
+        let mut children_map = HashMap::new();
+        children_map.insert(200, vec![201, 202]);
+        children_map.insert(300, vec![301]);
+
+        let mut pids = vec![100];
+        collect_descendants_from_map(100, &children_map, &mut pids);
+        assert_eq!(pids, vec![100]);
     }
 }
