+++ src/tui/home/tests.rs
@@ -2545,7 +2545,7 @@ fn test_apply_creation_results_returns_session_id() {
 fn test_project_group_name_uses_last_path_segment() {
     use super::project_group_name;
 
-    let mut inst = Instance::new("test", "/home/user/my-project");
+    let inst = Instance::new("test", "/home/user/my-project");
     assert_eq!(project_group_name(&inst), "my-project");
 }
