+++ tests/profile_management.rs
@@ -7,13 +7,8 @@ use agent_of_empires::session::{
 use anyhow::Result;
 use serial_test::serial;
 
-fn setup_temp_home() -> tempfile::TempDir {
-    let temp = tempfile::TempDir::new().unwrap();
-    std::env::set_var("HOME", temp.path());
-    #[cfg(target_os = "linux")]
-    std::env::set_var("XDG_CONFIG_HOME", temp.path().join(".config"));
-    temp
-}
+mod common;
+use common::setup_temp_home;
 
 #[test]
 #[serial]
